Introduce life-cycle-event-handler in order to toggle map#setMyLocationEnabled when app pauses / resumes.  This is important for other location-management modules which control location-services which desire to turn off location-services to conserve battery.  Also introduced a React bridge-module allowing the app code to execute methods upon the view instead of relying soley upon properties (eg: addLocation).  I've renamed the class RNGMapsModule to RNGMapsViewManager, since this class is not strictly a Module as it extends SimpleViewManager.  I've used the name RNGMapsModule then to implement the bridge-module.  Changed play-services dependency to import only com.google.android.gms:play-services-maps instead of including the entire, immense play-services library
