+++ android/app/build.gradle
@@ -23,6 +23,6 @@ dependencies {
   compile fileTree(dir: 'libs', include: ['*.jar'])
   testCompile 'junit:junit:4.12'
   compile 'com.android.support:appcompat-v7:23.1.0'
-  compile 'com.facebook.react:react-native:0.12.0'
-  compile 'com.google.android.gms:play-services:8.1.0'
+  compile 'com.facebook.react:react-native:0.14.+'
+  compile 'com.google.android.gms:play-services-maps:8.1.0'
 }
