+++ android/app/src/main/java/com/rota/rngmaps/RNGMapsPackage.java
@@ -15,15 +15,21 @@ import java.util.List;
  * Created by Henry on 08/10/2015.
  */
 public class RNGMapsPackage implements ReactPackage {
+    private RNGMapsViewManager viewManager;
+
+    public RNGMapsPackage() {
+        viewManager = new RNGMapsViewManager();
+    }
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactApplicationContext) {
-        return new ArrayList<>();
+        List<NativeModule> modules = new ArrayList<>();
+        modules.add(new RNGMapsModule(reactApplicationContext, viewManager));
+        return modules;
     }
-
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactApplicationContext) {
         return Arrays.<ViewManager>asList(
-                new RNGMapsModule()
+                viewManager
         );
     }
